Chop up GSUB subtables into their own set of files
