+++ src/opentype/tables/advanced/shared/subtables/lookuptypes/gsub-1.js
@@ -1,0 +1,8 @@
+import { LookupType } from "./gsub.js";
+
+export class LookupType1 extends LookupType {
+  constructor(p) {
+    super(p);
+    this.deltaGlyphID = p.int16;
+  }
+}
